Refactor to use sfml built in vector2 classes

+++ game_project/Point.cpp
@@ -1,7 +1,7 @@
 #include "Point.h"
 
 Point::Point(){}
-Point::Point(int x, int y) : x(x), y(y)
+Point::Point(float x, float y) : x(x), y(y)
 {
 }
